+++ tests/run_unittests.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
+echo "starting the database"
+service mariadb start
 
 apt-get update -qq
 apt-get install -y mariadb-client iputils-ping