+++ tests/run_webtests_basetest.sh
@@ -1,6 +1,5 @@
 #!/bin/bash
 
-echo "starting the database"
 service mariadb start
 
 apt-get update -qq