+++ CHANGELOG.md
@@ -6,10 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 - add `now` helper
+- ignore deactivated accounts in forgot app
 
 ## [6.25.1] - 2025-01-17
 - fix escape sequence deprecation warning in fill.py
-- ignore deactivated accounts in forgot app
 
 ## [6.25.0] - 2025-01-15
 - update zoom.tools.now() for compatibility with Python >= 3.11