+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - add `now` helper
 - ignore deactivated accounts in forgot app
 - use stable version of chrome in test pipelines
+- use prepared testbase image for pipeline testing
 
 ## [6.25.1] - 2025-01-17
 - fix escape sequence deprecation warning in fill.py