+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - use newer Site class from sites module by default
 - use stable version of chrome in test pipelines
 - make abs_url a request attribute
+- use prepared testbase image for pipeline testing
 
 ## [6.25.1] - 2025-01-17
 - fix escape sequence deprecation warning in fill.py