+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - use stable version of chrome in test pipelines
 - make abs_url a request attribute
 - use prepared testbase image for pipeline testing
+- add environment variable substitution in site.ini
 
 ## [6.25.1] - 2025-01-17
 - fix escape sequence deprecation warning in fill.py