+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [6.26.0] - 2025-01-17
 - add `now` helper
 - ignore deactivated accounts in forgot app
 - use newer Site class from sites module by default